Helper functions to get file names based on [basename]

+++ src/main/scala/edu/nyu/tandon/search/selective/data/results/trec/TrecLine.scala
@@ -3,11 +3,11 @@ package edu.nyu.tandon.search.selective.data.results.trec
 /**
   * @author michal.siedlaczek@nyu.edu
   */
-class TrecLine(val queryId: Int,
+class TrecLine(val queryId: Long,
                val topic: String,
                val title: String,
                val index: Int,
                val score: Double) {
-  def apply(queryId: Int, topic: String, title: String, index: Int, score: Double) = new TrecLine(queryId, topic, title, index, score)
+  def apply(queryId: Long, topic: String, title: String, index: Int, score: Double) = new TrecLine(queryId, topic, title, index, score)
   override def toString: String = s"$queryId\t$topic\t$title\t$index\t$score\tnull"
 }
